db: Set `rhodecode.is_test` in `make_pyramid_app` instead of `make_app`
Based on this setting the crypto backend is selected: brypt or md5.
During tests md5 is used to sppedup test execution. We have to move this
setting up to the pyramid layer otherwise the test DB init will use bcrypt
and the tests use md5. This will result in test users not able to login.