file-browser: refactor how we load metadata for file trees.
Before we used to use JSON data to map the nodes to json and fill in
metadata. Now we use rendered parts of html. This is nicer for caching as
it would allow us to replace the view with cached tree and then after ajax load
replace it again with cached with metadata. On the next request we'll
get the cached with metadata and thus we can skip entirely second
ajax call for metadata.

This is part of #4083