nix: Add backwardsCompatibleFetchgit
Needed so that we can use fetchgit with the branch nixos-16.03 as well as
current nixpkgs-unstable. It restores the behavior of nixos-16.03 which is
technically wrong since it removes all .git files.

Practically this issue has no big impact, so that we can keep it easily a few
months this way until we don't depend on building against nixos-16.03 anymore.