scm-app: allow to set documented pyro4 as scm_app_implementation.
The documented ini setting allows to set pyro4 as scm_app, while
the backend code crashes on import of pyro4 module. This change
allows to configure the app based on the .ini docs.

if we pass pyro4 as backedn, we simply skip import and use
the default pyro4 scm_app