repo: Implement ReadmeFinder
It currently resembles the logic from the old markup renderer. The
implementation is based on get_nodes though, so that we should usually only have
one call or at most three calls if the directories "doc" and/or "docs" exist.

A difference is that we now ignore the case, this already simplifies the rule
definition.